[REFACTOR] Modal code refactoring
refactoring modal code about pr massage
